backend/requirements.txt

# Framework Web
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Utilidades
python-dotenv>=1.0.0
pydantic>=2.8.0
pydantic-settings>=2.0.0

# Base de Datos
supabase>=2.3.0
postgrest>=0.16.0

# HTTP Client
httpx>=0.27.0

# LLM - Groq
groq>=0.4.0

# Embeddings y RAG
sentence-transformers>=2.3.0
numpy>=1.24.0

# Google Gemini - ACTUALIZADO
google-genai>=1.61.0

# File handling - NUEVO
python-multipart>=0.0.6
pillow>=10.0.0


# Observabilidad - Langfuse - NUEVO
langfuse>=2.6.0